Rename Math namespace to Mathematics to prevent naming collisions
